WIP
Added the modified script that needs a certain change to Eluna. The README inside the WIP folder has a link to the specific PR required, in case anyone wants to try it.

It adds the assistant as a regular vendor rather than using hardcoded items. It also adds containers, which was a feature previously removed. It does require the user to import the SQL file found in the WIP folder, but it's recommended to modify it before importing it if you want to make any changes.